+++ packages/mobile/App/ui/components/Forms/NewPatientForm/PatientPersonalInfoForm/KeyInformationSection.tsx
@@ -6,6 +6,7 @@ import { LocalisedField } from '~/ui/components/Forms/LocalisedField';
 import { GenderOptions, Gender } from '~/ui/helpers/constants';
 import { RadioButtonGroup } from '~/ui/components/RadioButtonGroup';
 import { DateField } from '~/ui/components/DateField/DateField';
+import { TextField } from '../../../TextField/TextField';
 
 export const KeyInformationSection = (): ReactElement => {
   const { getBool } = useLocalisation()
@@ -17,6 +18,7 @@ export const KeyInformationSection = (): ReactElement => {
     <FormGroup sectionName="KEY INFORMATION" marginTop>
       <LocalisedField name="sex" options={filteredGenderOptions} component={RadioButtonGroup} />
       <LocalisedField name="dateOfBirth" component={DateField} />
+      <LocalisedField name="email" component={TextField} />
     </FormGroup>
   );
 }