+++ packages/shared-src/src/models/Encounter.js
@@ -60,6 +60,7 @@ export class Encounter extends Model {
         'invoice',
         'invoice.invoiceLineItems',
         'invoice.invoicePriceChangeItems',
+        'documents',
       ],
       ...nestedSyncConfig,
       channelRoutes: [
@@ -158,7 +159,15 @@ export class Encounter extends Model {
   }
 
   static getFullReferenceAssociations() {
-    return ['vitals', 'department', 'location', 'examiner'];
+    return [
+      'vitals',
+      'department',
+      'examiner',
+      {
+        association: 'location',
+        include: ['Facility'],
+      },
+    ];
   }
 
   static initRelations(models) {
@@ -244,6 +253,11 @@ export class Encounter extends Model {
     this.hasMany(models.Triage, {
       foreignKey: 'encounterId',
       as: 'triages',
+    });
+
+    this.hasMany(models.DocumentMetadata, {
+      foreignKey: 'encounterId',
+      as: 'documents',
     });
 
     // this.hasMany(models.Procedure);