+++ packages/shared-src/src/tasks/CreateLabRequestNotifications.js
@@ -1,28 +1,6 @@
-import { Op } from 'sequelize';
 import config from 'config';
+import { getPatientSurveyResponseAnswer } from '../utils';
 import { ICAO_DOCUMENT_TYPES, LAB_REQUEST_STATUSES } from '../constants';
-
-async function fetchMostRecentEmailAddress(patientId, models) {
-  const questionId = config?.questionCodeIds?.email;
-  if (!questionId) {
-    return null;
-  }
-
-  // Find the most recent non-empty answer to the question
-  const answer = await models.SurveyResponseAnswer.findOne({
-    include: {
-      model: models.SurveyResponse,
-      as: 'surveyResponse',
-      include: { model: models.Encounter, where: { patientId }, as: 'encounter' },
-    },
-    where: {
-      data_element_id: questionId,
-      [Op.not]: [{ body: '' }],
-    },
-    order: [[{ model: models.SurveyResponse, as: 'surveyResponse' }, 'end_time', 'DESC']],
-  });
-  return answer?.body;
-}
 
 export async function createSingleLabRequestNotification(labRequest, models) {
   if (
@@ -31,6 +9,13 @@ export async function createSingleLabRequestNotification(labRequest, models) {
   ) {
     const encounter = await models.Encounter.findByPk(labRequest.encounterId);
 
+    const questionId = config?.questionCodeIds?.email;
+    const emailAddress = await getPatientSurveyResponseAnswer(
+      models,
+      encounter.patientId,
+      questionId,
+    );
+
     // Bulk create action sets off a hook to send these out
     await models.CertificateNotification.bulkCreate([
       {
@@ -38,7 +23,7 @@ export async function createSingleLabRequestNotification(labRequest, models) {
         requiresSigning: false,
         patientId: encounter.patientId,
         // If forward address is null, the communication service will attempt to use the patient.email field
-        forwardAddress: await fetchMostRecentEmailAddress(encounter.patientId, models),
+        forwardAddress: emailAddress,
         // TODO: attach lab test
       },
     ]);