+++ packages/sync-server/app/integrations/VdsNc/Document.js
@@ -1,13 +1,14 @@
 import { ICAO_DOCUMENT_TYPES } from 'shared/constants';
 import { log } from 'shared/services/logging';
-import { VdsNcSigner } from 'shared/models';
+import { Signer } from 'shared/models';
 import { depem, base64UrlEncode } from 'shared/utils';
-import { vdsConfig } from './Config';
 import { canonicalize } from 'json-canonicalize';
+import config from 'config';
 
 export class VdsNcDocument {
-  config = vdsConfig();
-  models = { VdsNcSigner };
+  config = config.integrations.signer;
+
+  models = { Signer };
 
   isSigned = false;
 
@@ -71,7 +72,7 @@ export class VdsNcDocument {
 
     log.debug('Signing VDS document');
 
-    const signer = await this.models.VdsNcSigner.findActive();
+    const signer = await this.models.Signer.findActive();
     if (!signer) throw new Error('No active signer');
     this.signer = signer;