+++ packages/sync-server/app/utils/makePatientCertificate.js
@@ -2,11 +2,23 @@ import React from 'react';
 import ReactPDF from '@react-pdf/renderer';
 import path from 'path';
 import QRCode from 'qrcode';
+import { get } from 'lodash';
+import config from 'config';
+
 import { log } from 'shared/services/logging';
-import { tmpdir, CovidLabCertificate, VaccineCertificate } from 'shared/utils';
-import { getLocalisationData } from './localisation';
+import {
+  tmpdir,
+  CovidLabCertificate,
+  VaccineCertificate,
+  getPatientSurveyResponseAnswer,
+} from 'shared/utils';
+import { generateUVCIForPatient } from '../integrations/VdsNc';
+import { getLocalisation } from '../localisation';
 
 export const makeVaccineCertificate = async (patient, printedBy, models, vdsData = null) => {
+  const localisation = await getLocalisation();
+  const getLocalisationData = key => get(localisation, key);
+
   const folder = await tmpdir();
   const fileName = `vaccine-certificate-${patient.id}.pdf`;
   const filePath = path.join(folder, fileName);
@@ -41,7 +53,7 @@ export const makeVaccineCertificate = async (patient, printedBy, models, vdsData
       include: models.PatientAdditionalData.getFullReferenceAssociations(),
     });
     const patientData = { ...patient.dataValues, additionalData: additionalData?.dataValues };
-    const uvci = await patient.getIcaoUVCI();
+    const uvci = await generateUVCIForPatient(patient.id);
 
     await ReactPDF.render(
       <VaccineCertificate
@@ -69,6 +81,9 @@ export const makeVaccineCertificate = async (patient, printedBy, models, vdsData
 };
 
 export const makeCovidTestCertificate = async (patient, printedBy, models, vdsData = null) => {
+  const localisation = await getLocalisation();
+  const getLocalisationData = key => get(localisation, key);
+
   const folder = await tmpdir();
   const fileName = `covid-test-certificate-${patient.id}.pdf`;
   const filePath = path.join(folder, fileName);
@@ -95,12 +110,41 @@ export const makeCovidTestCertificate = async (patient, printedBy, models, vdsDa
   });
 
   const vds = vdsData ? await QRCode.toDataURL(vdsData) : null;
+  const additionalData = await models.PatientAdditionalData.findOne({
+    where: { patientId: patient.id },
+    include: models.PatientAdditionalData.getFullReferenceAssociations(),
+  });
+  const passportFromSurveyResponse = await getPatientSurveyResponseAnswer(
+    models,
+    patient.id,
+    config?.questionCodeIds?.passport,
+  );
+
+  const nationalityId = await getPatientSurveyResponseAnswer(
+    models,
+    patient.id,
+    config?.questionCodeIds?.nationalityId,
+  );
+
+  const nationalityRecord = await models.ReferenceData.findByPk(nationalityId);
+  const nationalityFromSurveyResponse = nationalityRecord?.dataValues?.name;
+
+  const patientData = {
+    ...patient.dataValues,
+    additionalData: {
+      ...additionalData?.dataValues,
+      passport: additionalData?.dataValues?.passport || passportFromSurveyResponse,
+      nationality: {
+        name: additionalData?.dataValues?.nationality?.name || nationalityFromSurveyResponse,
+      },
+    },
+  };
 
   try {
-    const labs = await patient.getLabRequests();
+    const labs = await patient.getCovidLabTests();
     await ReactPDF.render(
       <CovidLabCertificate
-        patient={patient.dataValues}
+        patient={patientData}
         labs={labs}
         signingSrc={signingImage?.data}
         watermarkSrc={watermark?.data}