+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## vNEXT
 
 ### Manual steps required after upgrade ⚙
-
-- (left blank)
+- Add an entry for `questionCodeIds.passport` and `questionCodeIds.nationalityId` in the config to look up survey
+responses for patients in covid lab test certificates
 
 ### Features ⭐
 
@@ -13,6 +13,34 @@
 ### Tweaks ⚖️
 
 - (left blank)
+
+### Bug fixes 🐛
+
+- (left blank)
+
+### Infrastructure and maintenance 🛠
+
+- (left blank)
+
+## v1.14.0
+
+### Manual steps required after upgrade ⚙
+
+- Change `vaccineCertificateFooter` key in the config to `vaccineCertificate` refer to default.json
+- Consider adding `fiji-aspen-encounter-summary-line-list` to disabledReports
+- A server with invalid localisation will now exit: fix the localisation (preferred), or set `localisation.allowInvalidLocalisation` to `true` (if there's a good reason why it's invalid)
+- Add a `localisation.data.country` section to the config with country name and ISO 3166 codes
+- VDS Signer integration config structure has changed, refer to default.json
+
+### Features ⭐
+
+- Add new report `fiji-aspen-encounter-summary-line-list`
+
+### Tweaks ⚖️
+
+- Tweaked localisation to prepare for loading it from the db
+- "VDS" scheduled tasks were renamed to "Signer", `schedules` config keys have changed in consequence
+- Hid the "Forgot your password?" page and link
 
 ### Bug fixes 🐛